CM : Gestion de l'ajout et la modification d'une pizza

+++ src/main/java/com/tp/pizza/bll/PizzaService.java
@@ -7,7 +7,7 @@ public interface PizzaService {
 
   List<Pizza> getPizzas();
 
-  Pizza getPizzaById(int id);
+  Pizza getPizzaById(long id);
 
   void addPizza(Pizza pizza);
